config option to require the monitored path has a specific value
This allows for Victron Venus GX digital input pins to be used as a monitor path where there is always a SignalK path, but the value differs based on if the input is on or off.

In the case of the Victron Venus GX digital input number 1, the key is `electrical.venus-input.1.stateNumber`, it's value `11` for Stopped, and `10` for Running.  Alternatively, use `electrical.venus-input.1.state` with value `Running`.